requirements.txt
httpx>=0.25
python-socketio>=5.10
websocket-client>=1.6
requests>=2.31
python-dotenv>=1.0
